Add Node.js API
