+++ package.json
@@ -12,6 +12,7 @@
 	"bin": {
 		"wait-for-localhost": "cli.js"
 	},
+	"main": "index.js",
 	"engines": {
 		"node": ">=8"
 	},
@@ -19,7 +20,8 @@
 		"test": "xo && ava"
 	},
 	"files": [
-		"cli.js"
+		"cli.js",
+		"index.js"
 	],
 	"keywords": [
 		"cli-app",
