+++ test.js
@@ -2,8 +2,26 @@ import test from 'ava';
 import execa from 'execa';
 import delay from 'delay';
 import createTestServer from 'create-test-server';
+import waitForLocalhost from '.';
 
-test('main', async t => {
+test('api', async t => {
+	t.plan(2);
+
+	const server = await createTestServer();
+	server.head('/', async (request, response) => {
+		await delay(1000);
+		response.end();
+		t.pass();
+	});
+
+	await waitForLocalhost(server.port);
+
+	t.pass();
+
+	await server.close();
+});
+
+test('cli', async t => {
 	t.plan(2);
 
 	const server = await createTestServer();
